% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgesort.R
\name{edgesort}
\alias{edgesort}
\title{Edge Sort}
\usage{
edgesort(networkname)
}
\arguments{
\item{Name}{of a network created using bootnet}
}
\value{
A desecending list of edgeweights
}
\description{
This function uses the output from the 'estimateNetowrk' package of boonet to create sorted list (descending) of edge weights.
}
\examples{
network <- estimateNetwork(data,default="EPICglasso")
edgesort(network)
}
